backend/requirements.txt
fastapi==0.111.0
requests==2.32.3
uvicorn==0.30.1
groq==0.30.0
pydantic-settings==2.3.0
beautifulsoup4
selenium
python-dotenv
pymongo[srv]